+++ adapters/Fyber/CHANGELOG.md
@@ -1,5 +1,8 @@
 ## Fyber iOS Mediation Adapter Changelog
 
+#### Next Version
+- Adapter will now initialize the Fyber SDK before making an ad request if the Fyber SDK has not been initialized yet.
+
 #### Version 7.4.0.0
 - Initial release!
 - Added support for banner, interstitial and rewarded ad formats.